backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
vaderSentiment==3.3.2
spotipy==2.23.0
python-dotenv==1.0.0
cachetools==5.3.2
pydantic==2.5.2